# Optimize marshalTags()